+++ Tests/KeyboardKitTests/Layout/Providers/SystemKeyboardLayoutProviderTests.swift
@@ -39,8 +39,8 @@ class SystemKeyboardLayoutProviderTests: QuickSpec {
                 let actions = provider.actions(for: context, inputs: inputs)
                 let items = provider.items(for: context, actions: actions)
                 let layout = provider.keyboardLayout(for: context)
-                let expected = KeyboardLayout(items: items)
-                expect(layout.items.count).to(equal(expected.items.count))
+                let expected = KeyboardLayout(itemRows: items)
+                expect(layout.itemRows.count).to(equal(expected.itemRows.count))
             }
         }